+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### BREAKING CHANGES
 - The codebase has migrated to TypeScript
 
+
 ## [4.1.0]
 
 ### Pre-built version